+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestTask.groovy
@@ -163,11 +163,16 @@ class PitestTask extends JavaExec {
     @Input
     FileCollection launchClasspath
 
+    @Input
+    @Optional
+    Map<String, String> pluginConfiguration
+
     @Override
     void exec() {
         Map<String, String> taskArgumentsMap = createTaskArgumentMap()
-        def argsAsList = createArgumentsListFromMap(taskArgumentsMap)
-        setArgs(argsAsList)
+        List<String> argsAsList = createArgumentsListFromMap(taskArgumentsMap)
+        List<String> multiValueArgsAsList = createMultiValueArgsAsList()
+        setArgs(argsAsList + multiValueArgsAsList)
         setMain("org.pitest.mutationtest.commandline.MutationCoverageReport")
         setJvmArgs(getMainProcessJvmArgs() ?: getJvmArgs())
         setClasspath(getLaunchClasspath())
@@ -235,5 +240,15 @@ class PitestTask extends JavaExec {
             argList.add("--" + k + "=" + v)
         }
         argList
+    }
+
+    @VisibleForTesting
+    List<String> createMultiValueArgsAsList() {
+        //It is a duplication/special case handling, but a PoC implementation with emulated multimap was also quite ugly and in addition error prone
+        getPluginConfiguration()?.collect {
+            "${it.key}=${it.value}".toString()
+        }?.collect {
+            "--pluginConfiguration=$it"
+        } ?: []
     }
 }